Bump deps, add test util package
