+++ examples/parcel/package.json
@@ -8,12 +8,12 @@
   "dependencies": {
     "@mdx-js/react": "^1.1.6",
     "@mdx-js/tag": "^0.20.3",
-    "react": "^0.14.x || ^15.x || ^16.x",
-    "react-dom": "^16.8.6"
+    "react": "^16.x",
+    "react-dom": "^16.9.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
     "@mdx-js/parcel-plugin-mdx": "^1.1.6",
     "parcel-bundler": "^1.12.3"
