+++ packages/gatsby-theme-mdx/package.json
@@ -9,12 +9,12 @@
     "clean": "gatsby clean"
   },
   "devDependencies": {
-    "gatsby": "2.13.45"
+    "gatsby": "2.13.53"
   },
   "dependencies": {
     "@blocks/kit": "0.0.28",
-    "@emotion/core": "10.0.14",
-    "@emotion/styled": "10.0.14",
+    "@emotion/core": "10.0.15",
+    "@emotion/styled": "10.0.15",
     "@mdx-js/mdx": "^1.1.6",
     "@mdx-js/react": "^1.1.6",
     "@mdx-js/runtime": "^1.1.6",
@@ -22,16 +22,16 @@
     "@reach/skip-nav": "0.1.3",
     "@rebass/emotion": "3.2.0",
     "emotion-theming": "10.0.14",
-    "gatsby": "2.13.45",
+    "gatsby": "2.13.53",
     "gatsby-plugin-catch-links": "2.1.2",
     "gatsby-plugin-compile-es6-packages": "2.1.0",
     "gatsby-plugin-emotion": "4.1.2",
     "gatsby-plugin-fathom": "1.1.0",
     "gatsby-plugin-google-fonts": "1.0.1",
-    "gatsby-plugin-mdx": "1.0.19",
+    "gatsby-plugin-mdx": "1.0.23",
     "gatsby-plugin-page-creator": "2.1.5",
-    "gatsby-plugin-react-helmet": "3.1.2",
-    "gatsby-plugin-theme-ui": "0.2.26",
+    "gatsby-plugin-react-helmet": "3.1.3",
+    "gatsby-plugin-theme-ui": "0.2.29",
     "is-absolute-url": "3.0.0",
     "lodash.flatten": "4.4.0",
     "prism-react-renderer": "0.1.7",
@@ -41,10 +41,10 @@
     "remark-mdx-remove-exports": "^1.1.6",
     "remark-mdx-remove-imports": "^1.1.6",
     "remark-slug": "5.1.2",
-    "theme-ui": "0.2.26"
+    "theme-ui": "0.2.31"
   },
   "peerDependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0"
   }
 }
