+++ packages/react/package.json
@@ -32,12 +32,12 @@
     "@babel/preset-react": "7.0.0",
     "jest": "24.8.0",
     "microbundle": "0.11.0",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
     "rimraf": "2.6.3"
   },
   "peerDependencies": {
-    "react": "^16.8.0"
+    "react": "^16.9.0"
   },
   "scripts": {
     "build": "microbundle --format es,cjs --external react --jsx React.createElement --no-sourcemap --no-compress",
