+++ packages/remark-mdx/package.json
@@ -34,7 +34,7 @@
     "@babel/plugin-syntax-jsx": "7.2.0",
     "@mdx-js/util": "^1.1.6",
     "is-alphabetical": "1.0.3",
-    "remark-parse": "7.0.0",
+    "remark-parse": "7.0.1",
     "unified": "8.3.2"
   },
   "scripts": {
@@ -45,7 +45,7 @@
   },
   "devDependencies": {
     "jest": "24.8.0",
-    "remark-stringify": "7.0.1",
+    "remark-stringify": "7.0.2",
     "vfile": "4.0.1"
   }
 }
