+++ packages/runtime/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "@mdx-js/mdx": "^1.1.6",
     "@mdx-js/react": "^1.1.6",
-    "buble": "^0.19.6"
+    "buble": "^0.19.8"
   },
   "devDependencies": {
     "@babel/core": "7.5.5",
@@ -44,14 +44,14 @@
     "jest": "24.8.0",
     "microbundle": "0.11.0",
     "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react-dom": "16.9.0",
     "rehype-add-classes": "1.0.0",
     "remark-autolink-headings": "5.2.1",
     "remark-slug": "5.1.2",
     "rimraf": "2.6.3"
   },
   "peerDependencies": {
-    "react": "^16.8.0"
+    "react": "^16.9.0"
   },
   "jest": {
     "testEnvironment": "node"
