+++ packages/test-util/readme.md
@@ -1,0 +1,15 @@
+# [@mdx-js/test-util][mdx]
+
+Test utility functions used in [MDX][].
+
+## License
+
+[MIT][] © [John Otander][author]
+
+<!-- Definitions -->
+
+[mit]: license
+
+[mdx]: https://github.com/mdx-js/mdx
+
+[author]: https://johno.com
