+++ packages/vue-loader/package.json
@@ -40,7 +40,7 @@
     "babel-plugin-transform-vue-jsx": "4.0.1",
     "jest": "24.8.0",
     "memory-fs": "0.4.1",
-    "react": "16.8.6",
+    "react": "16.9.0",
     "webpack": "4.39.1"
   },
   "scripts": {
